abuse/signup: Prevent online CAPTCHA bruteforcing
Many of the CAPTCHAs produced by TextCaptcha contain either a list of
words (possible answers) or have an answer that is a small (<100) number.

As such, requiring that the user finds a solution within a given amount
of time, with access to an online verifier (the API) is not enough:
instead, we restrict them to finding the answer *on the first try*.

The trade-off there is that we must keep track of transient
state (whether a not-yet-expired token was already submitted with a
wrong answer).